# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NS Scrap ERP Next API
  version: 0.1.0
  description: >
    Baseline API catalog for the active Next.js app. This file records current
    route handlers first; detailed request and response schemas should be added
    as each module batch hardens its contract.
servers:
  - url: http://localhost:3000
    description: Local Next.js development server
  - url: https://new-ns-scrap-erp.vercel.app
    description: Current Vercel deployment, if available
security:
  - SupabaseSession: []
...
paths:
...
  /api/stock/status-convert:
    get:
      tags: [Stock]
      operationId: listStockStatusConverts
      summary: List stock status convert history
      responses:
        "200":
          description: Status convert rows and stock reference options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockOperationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Stock]
      operationId: createStockStatusConvert
      summary: Convert product stock status between RM, WIP, and FG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusConvertRequest"
      responses:
        "200":
          $ref: "#/components/responses/StockWriteOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase Auth session cookie or equivalent authenticated request context.
...
  responses:
...
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Supabase session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    StockWriteOk:
      description: Stock write completed and traceable ledger rows were created
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StockWriteResponse"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      additionalProperties: true
    OptionalDocNo:
      type: [string, "null"]
      maxLength: 40
      pattern: "^[A-Za-z0-9_-]+$"
...
    StockBaseWriteRequest:
      type: object
      required: [branchId, date, productId, warehouseId]
      properties:
        branchId:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        docNo:
          $ref: "#/components/schemas/OptionalDocNo"
        lotNo:
          type: [string, "null"]
          maxLength: 80
        notes:
          type: [string, "null"]
          maxLength: 500
        productId:
          type: string
          minLength: 1
        warehouseId:
          type: string
          minLength: 1
...
    StockOption:
      type: object
      required: [id, name]
      properties:
        active:
          type: [boolean, "null"]
        branchId:
          type: [string, "null"]
        code:
          type: [string, "null"]
        id:
          type: string
        name:
          type: string
        status:
          type: [string, "null"]
    StockOperationListResponse:
      type: object
      required: [reference, rows]
      properties:
        reference:
          type: object
          required: [branches, products, warehouses]
          properties:
            branches:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
            customers:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
            products:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
            warehouses:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
    StockStatus:
      type: string
      enum: [RM, WIP, FG]
    StockWriteResponse:
      type: object
      required: [ok]
      properties:
        id:
          type: string
        ok:
          type: boolean
        refId:
          type: string
        refNo:
          type: string
    StatusConvertRequest:
      allOf:
        - $ref: "#/components/schemas/StockBaseWriteRequest"
        - type: object
          required: [fromStatus, qty, toStatus]
          properties:
            fromStatus:
              $ref: "#/components/schemas/StockStatus"
            qty:
              type: number
              exclusiveMinimum: 0
            reason:
              type: [string, "null"]
              maxLength: 240
            toStatus:
              $ref: "#/components/schemas/StockStatus"